feat: better config
We now can set the sequence element with :config/sequence-element. This does mean we have to be explicit about translators with :config/translators. For backwards compatibility a top level translator map is still supported.

HOWEVER: *feature-selector* removed because the default was a bit ackward (it was the first nested content). This means we have to fix all current configs. Because of that we could just remove backwards compatibility for translators