fix: multi element support (#10)
The old solution returned an array if multiple elements and just an object if only on. This does not work because it cannot return single sized arrays.

This new solution adds an hashset to xml2json/mapped for :arrays. If the key is in the set the resulting value is a array. For example BAG VBO has :arrays #{:gebruiksdoelVerblijfsobject} in #xml2json/mapped. So it returns single elements if not an array (the default) an otherwise always an array